+++ test/supports/app_run_tests/app_was_once_stub.py
@@ -1,9 +1,0 @@
-# Copyright Modal Labs 2024
-import modal
-
-app = modal.Stub()
-
-
-@app.function()
-def foo():
-    print("foo")